Continued thinking out loud about the role of Lua in the LuaService, including a plan for supporting more than one installation of LuaService on a single PC.
Cleaned up and documented LuaMain.c, including steps in the direction of a real interface between the service framework and the Lua code.

Made the project link against LUA5.1.LIB, so it has a dependence on the LUA5.1.DLL from Lua Binaries. This will make use of other modules that assume that naming convention easier.

Demonstrated the debug build running as an installed service using service.print and service.sleep from a test lua script.