Snapshot of a runnable build. The script still has a constant name and all of the proper configuration is yet to be done.
A nasty bug in initGlobals() is fixed that messed up the definitions of service.path and service.filename.

A nasty bug in service.print() is fixed that ate the stack alive while constructing the string buffer to pass to OutputDebugSring().

The project is modified to include the Lua library as an Eclipse linked resource based on an Eclipse variable on the premise that this will make the configuration differences between developer machines hide behind a per-machine definition of the variable LuaRoot.