Changed all the functions that take a HANDLE as the first argument to be methods of a new class HPORT.
Created the new class HPORT with constructors and destructor that can
open a named port and close the port on destruction.

Added test cases to check for HPORT creation and destruction.

Changed all the simple method test cases to use the new mechanim.

Brought the TODO list in readme.txt up to date.